moved markdown parser and serializer into their own file
